+++ @MIC_DCAM4Camera/old/setGetProperty.m
