+++ source/DCAM4/DCAM4CopyFrameBundle/DCAM4CopyFrameBundle.vcxproj
@@ -19,6 +19,7 @@
     </ProjectConfiguration>
   </ItemGroup>
   <ItemGroup>
+    <ClCompile Include="..\share\helper.cpp" />
     <ClCompile Include="mexFunction.cpp" />
   </ItemGroup>
   <ItemGroup>
