+++ source/DCAM4/DCAM4QueryValue/matlab.def
@@ -1,0 +1,2 @@
+LIBRARY
+	EXPORTS mexFunction
