+++ source/DCAM4/share/helper.h
@@ -1,0 +1,1 @@
+BOOL get_framebundle_information(HDCAM hdcam, int32& number_of_bundle, int32& width, int32& height, int32& rowbytes, int32& totalframebytes, int32& framestepbytes);
